refactor : 해당 클래스들 post/application 폴더 밑으로 이동
