+++ src/main/java/com/kakaotech/team14backend/post/application/GetMyPostUsecase.java
@@ -1,8 +1,7 @@
 package com.kakaotech.team14backend.post.application;
 
-import com.kakaotech.team14backend.post.application.FindMyPostUsecase;
-import com.kakaotech.team14backend.outer.post.dto.GetMyPostResponseDTO;
-import com.kakaotech.team14backend.outer.post.dto.GetPostDTO;
+import com.kakaotech.team14backend.post.dto.GetMyPostResponseDTO;
+import com.kakaotech.team14backend.post.dto.GetPostDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
