+++ src/main/java/com/kakaotech/team14backend/post/application/UpdatePostLikeCountUsecase.java
@@ -1,8 +1,8 @@
 package com.kakaotech.team14backend.post.application;
 
 import com.kakaotech.team14backend.post.domain.PostLikeCount;
+import com.kakaotech.team14backend.post.dto.GetPostLikeCountDTO;
 import com.kakaotech.team14backend.post.infrastructure.PostLikeCountRepository;
-import com.kakaotech.team14backend.outer.post.dto.GetPostLikeCountDTO;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Component;
